Properly remove the scroll event handler when the page doesn’t have a data-scrollinity-path element, using a namespaced jquery event. (See // https://css-tricks.com/namespaced-events-jquery/)
It is necessary to manually garbage-collect event handlers that are bound to top level objects such as $(window) when navigating pages with Ajax (such as with Turbolinks). If the event handler is not removed, then when the user scrolls to the bottom of the next page, it will continue to call the scrolling event handler in the background!